tuf Updater: fix snapshot version rollback case
The snapshot version listed in timestamp must never decrease
(except if timestamp keys are rotated but that in that case local timestamp
is not used at all).

Theoretically this is security relevant for a TUF client but for a sigstore
TUF client it really isn't: Timestamp/snapshot keys are exposed in the same
environments so compromise in one means compromise for both (because of this
they are in fact the same key). So tricks like snapshot version rollback
are really not that relevant.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>